Move time functionality into platform.c
